Remove comments from config file | remove useless resolvers

+++ src/graphql/resolvers/user_resolver.js
@@ -15,12 +15,6 @@ const UserResolver = {
       try {
         return User.findById(_id)
       } catch (error) { throw new Error(error) }
-    },
-    token: async (root, { token }) => {
-      if(!token) { throw new Error('Problemas na identificação do token') }
-      try {
-        return User.checkToken(token)
-      } catch (error) {throw new Error(error)}
     }
   },
   Mutation: {
@@ -37,6 +31,7 @@ const UserResolver = {
       } catch (error) { throw new Error(error) }
     },
     delete: async (root, args, { user }) => {
+      if (!user) { throw new Error('Faça login primeiro') }
       try {
         return User.findByIdAndRemove(user._id)
       } catch (error) { throw new Error(error) }
